+++ lib/flat_pack.rb
@@ -6,6 +6,8 @@ require "flat_pack/attribute_sanitizer"
 require "flat_pack/install_contract"
 require "flat_pack/install_verifier"
 require "flat_pack/rich_text_sanitizer"
+require "flat_pack/tiptap"
+require "flat_pack/token_auditor"
 
 module FlatPack
   HEROICON_VARIANTS = %i[outline solid mini micro].freeze